correction int -> float

+++ src/cercle.cpp
@@ -3,13 +3,13 @@
 
 void cercle::surface(float rayon)
 {
-int surf= (rayon*rayon)*3.14;
+float surf= (rayon*rayon)*3.14;
 
   std::cout << "l'aire du cercle est " << surf << std::endl;
 }
 
 void cercle::perimetre(float rayon)
 {
-int peri= 2*rayon*3.14;
+float peri= 2*rayon*3.14;
   std::cout << "le perimetre du cercle est " << peri << std::endl;
 }
